refactor(Winning): 당첨 개수 파라미터 타입 변경
int에서 long으로 변경

+++ src/main/java/lotto/model/Winning.java
@@ -29,7 +29,7 @@ public enum Winning {
 				.collect(Collectors.toMap(winning -> winning, winning -> INITIAL_VALUE));
 	}
 
-	public static Winning getWinningResult(int count, boolean hasBonusNumber) {
+	public static Winning getWinningResult(long count, boolean hasBonusNumber) {
 		Winning result = Arrays.stream(Winning.values())
 				.filter(winning -> winning.count == count)
 				.findFirst()
